Add Test for Add Export and verify existing test
