+++ vite.config.ts
@@ -84,7 +84,7 @@ export default defineConfig({
     port: 4000,
     proxy: {
       "/api": {
-        target: process.env.CARE_API ?? "https://careapi.ohc.network",
+        target: process.env.CARE_API ?? "http://192.168.0.204:9000/",
         changeOrigin: true,
       },
     },
@@ -93,7 +93,7 @@ export default defineConfig({
     port: 4000,
     proxy: {
       "/api": {
-        target: process.env.CARE_API ?? "https://careapi.ohc.network",
+        target: process.env.CARE_API ?? "http://192.168.0.204:9000/",
         changeOrigin: true,
       },
     },
